+++ scripts/lichessTools-init.js
@@ -101,6 +101,7 @@ lichessTools.loadTool(LiChessTools.Tools.FullScreen);
 lichessTools.loadTool(LiChessTools.Tools.OneClickMove);
 lichessTools.loadTool(LiChessTools.Tools.CommonTeams);
 lichessTools.loadTool(LiChessTools.Tools.ExpandAll);
+lichessTools.loadTool(LiChessTools.Tools.StudyLinks);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);